+++ lib/hooks/useGetOrgRepos.ts
@@ -6,18 +6,16 @@ const baseUrl = "https://api.github.com";
 const githubApiRepoFetcher: Fetcher = async (apiUrl: string) => {
   const sessionResponse = await supabase.auth.getSession();
   const sessionToken = sessionResponse?.data.session?.provider_token;
-  const defaultHeaders = {
+  const headers = {
     accept: "application/json",
+    ...(sessionToken
+      ? {
+          Authorization: `Bearer ${sessionToken}`,
+        }
+      : {}),
   };
   const res = await fetch(`${baseUrl}/${apiUrl}`, {
-    headers: {
-      ...defaultHeaders,
-      ...(sessionToken
-        ? {
-            Authorization: `Bearer ${sessionToken}`,
-          }
-        : {}),
-    },
+    headers,
   });
 
   if (!res.ok) {
@@ -31,12 +29,41 @@ const githubApiRepoFetcher: Fetcher = async (apiUrl: string) => {
     throw error;
   }
 
-  return res.json();
+  const linkHeader = res.headers.get("link");
+
+  if (linkHeader && linkHeader.includes('rel="last"')) {
+    // @ts-expect-error the regex group will exist as the link header is present as per https://docs.github.com/en/rest/using-the-rest-api/using-pagination-in-the-rest-api?apiVersion=2022-11-28#using-link-headers
+    const groups = /page=(?<lastPage>\d+)>;\s+rel="last"/.exec(linkHeader).groups as { lastPage: string };
+
+    const lastPage = Number(groups.lastPage);
+    // incrementing by two as we already have the first response for the first page of data
+    const pageNumbers = Array.from({ length: lastPage }, (_, i) => i + 2);
+    const pageUrls = pageNumbers.map((page) => {
+      const url = new URL(`${baseUrl}/${apiUrl}`);
+      url.searchParams.set("page", page.toString());
+      return url.toString();
+    });
+
+    const responses = await Promise.all(
+      pageUrls.map((url) =>
+        fetch(url, {
+          headers,
+        })
+      )
+    );
+
+    // add the initial request's reponse to the start of the array of paged data responses
+    responses.unshift(res);
+
+    const data = await Promise.all(responses.map((response) => response.json()));
+
+    return data.flat();
+  } else {
+    return res.json();
+  }
 };
 
 export const useGetOrgRepos = ({ organization, limit = 100 }: { organization: string | undefined; limit?: number }) => {
-  // TODO: get private repos appearing in the list
-  // See https://docs.github.com/en/rest/repos/repos for more information
   const query = new URLSearchParams();
   query.set("per_page", `${limit}`);
   query.set("type", "public");