+++ pages/workspaces/[workspaceId]/settings.tsx
@@ -25,6 +25,7 @@ import { WORKSPACE_UPDATED_EVENT } from "components/shared/AppSidebar/AppSidebar
 import { WorkspacesTabList } from "components/Workspaces/WorkspacesTabList";
 import { useGetWorkspaceContributors } from "lib/hooks/api/useGetWorkspaceContributors";
 import { TrackedContributorsTable } from "components/Workspaces/TrackedContributorsTable";
+import { deleteCookie } from "lib/utils/server/cookies";
 
 const DeleteWorkspaceModal = dynamic(() => import("components/Workspaces/DeleteWorkspaceModal"), { ssr: false });
 
@@ -46,6 +47,8 @@ export const getServerSideProps = async (context: GetServerSidePropsContext) => 
   });
 
   if (error) {
+    deleteCookie(context.res, "workspaceId");
+
     if (error.status === 404) {
       return { notFound: true };
     }